VPP: Initialize packet metadata fields to 0 (#26359)
- vnet_buffer and vnet_buffer2 areas of vpp packet buffer do not seem to
  be zeroed for every packet. While debugging in QA setup seeing some
  scenarios where there are leftover values from previous packets
  causing incorrect flow programming.

- Change rx/tx xlate_id to 32 bit in session_info_t

+++ nic/vpp/impl/sess.h
@@ -14,10 +14,10 @@ extern "C" {
 
 typedef struct session_info_s {
     uint16_t tx_rewrite_flags;
-    uint16_t tx_xlate_id;
+    uint32_t tx_xlate_id;
     uint16_t tx_xlate_id2;
     uint16_t rx_rewrite_flags;
-    uint16_t rx_xlate_id;
+    uint32_t rx_xlate_id;
     uint16_t rx_xlate_id2;
     uint16_t meter_id;
     uint64_t timestamp;
